Mock Demo
-Added GestureListener and DeviceContextData logging to login view
rather than StartActivity
