+++ app/src/main/java/group1/cpsc319/plurilock_client/Presenter/LoginActivity.java
@@ -4,6 +4,9 @@ import android.app.Activity;
 import android.content.Context;
 import android.content.Intent;
 import android.os.Bundle;
+import android.util.Log;
+import android.view.GestureDetector;
+import android.view.MotionEvent;
 import android.view.View;
 import android.widget.Button;
 
@@ -13,6 +16,8 @@ import group1.cpsc319.plurilock_client.R;
  * Created by anneunjungkim on 2016-02-27.
  */
 public class LoginActivity extends Activity {
+
+    private GestureDetector mGestureDetector;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -30,5 +35,82 @@ public class LoginActivity extends Activity {
                 startActivity(intent);
             }
         });
+
+        // Grab Device Specific Info at Startup
+        GeoInfoRetriever geoInfo = new GeoInfoRetriever();
+        HardwareInfoRetriever hardwareInfo = new HardwareInfoRetriever(this);
+
+        geoInfo.logDeviceInfo();
+        hardwareInfo.logDeviceInfo();
+
+        // Create an object of our Custom Gesture Detector Class
+        CustomGestureDetector customGestureDetector = new CustomGestureDetector();
+        // Create a GestureDetector
+        mGestureDetector = new GestureDetector(this, customGestureDetector);
+        // Attach listeners that'll be called for double-tap and related gestures
+        mGestureDetector.setOnDoubleTapListener(customGestureDetector);
+    }
+
+    @Override
+    public boolean onTouchEvent(MotionEvent event) {
+        mGestureDetector.onTouchEvent(event);
+
+        return super.onTouchEvent(event);
+    }
+
+    class CustomGestureDetector implements GestureDetector.OnGestureListener,
+            GestureDetector.OnDoubleTapListener {
+
+        @Override
+        public boolean onSingleTapConfirmed(MotionEvent e) {
+            Log.d(StartActivity.class.getSimpleName(), "onSingleTapConfirmed");
+            return true;
+        }
+
+        @Override
+        public boolean onDoubleTap(MotionEvent e) {
+            Log.d(StartActivity.class.getSimpleName(), "onDoubleTap");
+            return true;
+        }
+
+        @Override
+        public boolean onDoubleTapEvent(MotionEvent e) {
+            Log.d(StartActivity.class.getSimpleName(), "onDoubleTapEvent");
+            return true;
+        }
+
+        @Override
+        public boolean onDown(MotionEvent e) {
+            Log.d(StartActivity.class.getSimpleName(), "onDown");
+            return true;
+        }
+
+        @Override
+        public void onShowPress(MotionEvent e) {
+            Log.d(StartActivity.class.getSimpleName(), "onShowPress");
+        }
+
+        @Override
+        public boolean onSingleTapUp(MotionEvent e) {
+            Log.d(StartActivity.class.getSimpleName(), "onSingleTapUp");
+            return true;
+        }
+
+        @Override
+        public boolean onScroll(MotionEvent e1, MotionEvent e2, float distanceX, float distanceY) {
+            Log.d(StartActivity.class.getSimpleName(), "onScroll");
+            return true;
+        }
+
+        @Override
+        public void onLongPress(MotionEvent e) {
+            Log.d(StartActivity.class.getSimpleName(), "onLongPress");
+        }
+
+        @Override
+        public boolean onFling(MotionEvent e1, MotionEvent e2, float velocityX, float velocityY) {
+            Log.d(StartActivity.class.getSimpleName(), "onFling");
+            return true;
+        }
     }
 }
